+++ batch_4/src/integrations/registry/index.ts
@@ -1,8 +1,0 @@
-export {
-  createProviderRegistry,
-  getProviderRegistry,
-  resetProviderRegistry,
-  type ProviderRegistry,
-  type ProviderRegistryEntry,
-  type IntegrationProvider,
-} from './providerRegistry'